split definitions into separate files, mainly one type per file
