+++ a_doc.go
@@ -1,0 +1,11 @@
+// Copyright 2017 The oksvg Authors. All rights reserved.
+//
+// created: 2/12/2017 by S.R.Wiley
+// The oksvg package provides a partial implementation of the SVG 2.0 standard.
+// It can perform all SVG2.0 path commands, including arc and miterclip. It also
+// has some additional capabilities like arc-clip. Svgdraw does
+// not implement all SVG features such as animation or markers, but it can draw
+// the many of open source SVG icons correctly. See Readme for
+// a list of features.
+
+package oksvg
